# docs/api/postman/whusnet-portal-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Whusnet Operasional API",
        "version": "0.0.1",
...
    },
    "servers": [
        {
            "url": "https://safari-moon-profession-comedy.trycloudflare.com"
        }
    ],
...
    "paths": {
...
        "/api/customer-portal/me": {
            "get": {
                "operationId": "customer-portal.me.show",
                "description": "Nama, status layanan, paket aktif, alamat generic (desa/kecamatan \u2014\nbukan alamat detail/koordinat).",
                "summary": "Profil ringkas",
                "tags": [
                    "Portal Pelanggan"
                ],
                "responses": {
                    "200": {
                        "description": "Profil berhasil diambil.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "examples": [
                                        {
                                            "data": {
                                                "login_id": "PNG-RQ000631",
                                                "full_name": "Budi Santoso",
                                                "status": "active",
                                                "package": "Home 20 Mbps",
                                                "village": "Joresan",
                                                "district": "Mlarak",
                                                "claimed_at": "2026-08-01T10:00:00+07:00"
                                            },
                                            "meta": {
                                                "generated_at": "2026-08-25T09:00:00+07:00"
                                            }
                                        }
                                    ],
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CustomerPortalProfileResource"
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "generated_at": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "generated_at"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "description": "Access token tidak disertakan atau tidak valid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "PortalClientSecret": [],
                        "PortalAccessToken": []
                    }
                ]
            }
        },
        "/api/customer-portal/me/password": {
            "put": {
                "operationId": "customer-portal.me.update-password",
                "description": "`current_password` wajib. Sukses mencabut semua token LAIN (sesi\npemanggil tetap hidup).",
                "summary": "Ganti password",
                "tags": [
                    "Portal Pelanggan"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/PortalUpdatePasswordRequest"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "current_password": {
                                                "type": "string",
                                                "description": "Password saat ini \u2014 wajib, sesi yang dicuri tidak cukup untuk mengambil alih akun permanen.",
                                                "examples": [
                                                    "Kuda-Nil-Rajin-88"
                                                ]
                                            },
                                            "new_password": {
                                                "type": "string",
                                                "description": "Password baru, minimal 10 karakter, tidak boleh mengandung login_id/nomor HP, tidak boleh dari daftar password umum.",
                                                "examples": [
                                                    "Layang-Layang-Sore-42"
                                                ],
                                                "minLength": 10
                                            }
                                        },
                                        "required": [
                                            "current_password",
                                            "new_password"
                                        ]
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password berhasil diganti.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "examples": [
                                        {
                                            "message": "Password berhasil diganti."
                                        }
                                    ],
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "Password berhasil diganti."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Password saat ini salah, atau new_password tidak memenuhi syarat (< 10 karakter, mengandung login_id/nomor HP, atau daftar password umum).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "Password saat ini salah."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "description": "Access token tidak disertakan atau tidak valid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Jumlah percobaan melebihi batas (limiter sama seperti login).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "PortalClientSecret": [],
                        "PortalAccessToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "PopDistribusiReadToken": {
                "type": "http",
                "description": "Token akses baca-saja untuk membaca struktur topologi Mini POP dan Distribusi.",
                "scheme": "bearer"
            },
            "NetworkAssignmentWriteToken": {
                "type": "http",
                "description": "Token akses tulis untuk mengonfirmasi penugasan jaringan pelanggan.",
                "scheme": "bearer"
            },
            "PortalClientSecret": {
                "type": "apiKey",
                "description": "Client secret portal, statis per environment \u2014 membuktikan \"ini portal resmi\". Dicek di SEMUA endpoint /api/customer-portal/auth/* dan /me/*, TIDAK di /ping.",
                "in": "header",
                "name": "X-Portal-Client"
            },
            "PortalAccessToken": {
                "type": "http",
                "description": "Access token pelanggan (15 menit, dari /auth/login atau /auth/refresh) \u2014 membuktikan \"ini pelanggan X\". TIDAK dipakai di /auth/login, /auth/claim, /auth/refresh (belum ada identitas pelanggan saat endpoint itu dipanggil).",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "CustomerPortalProfileResource": {
                "type": "object",
                "properties": {
                    "login_id": {
                        "type": "string"
                    },
                    "full_name": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "package": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "village": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "district": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "claimed_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "login_id",
                    "full_name",
                    "status",
                    "package",
                    "village",
                    "district",
                    "claimed_at"
                ],
                "title": "CustomerPortalProfileResource"
            },
...
            "PortalUpdatePasswordRequest": {
                "type": "object",
                "properties": {
                    "current_password": {
                        "type": "string"
                    },
                    "new_password": {
                        "type": "string",
                        "minLength": 10
                    }
                },
                "required": [
                    "current_password",
                    "new_password"
                ],
                "title": "PortalUpdatePasswordRequest"
            },
...
        },
        "responses": {
...
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}